Récupérations et ajout des bd auth + scheduler

+++ src/main/java/com/example/sns/model/recup/bank1/ExternalTransactionsModel.java
@@ -15,7 +15,7 @@ public class ExternalTransactionsModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "external_transaction_id")
-    private int externalTransactionId;
+    private Long externalTransactionId;
 
     @Column(name = "local_transaction_id", nullable = false)
     private int localTransactionId;
