+++ src/main/java/com/example/sns/model/recup/bank2/ExternalTransactionsModel2.java
@@ -15,7 +15,7 @@ public class ExternalTransactionsModel2 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "external_transaction_id")
-    private int externalTransactionId;
+    private Long externalTransactionId;
 
     @Column(name = "local_transaction_id", nullable = false)
     private int localTransactionId;
