+++ src/main/java/com/example/sns/projection/bank2/AccountsProjection2.java
@@ -3,7 +3,7 @@ package com.example.sns.projection.bank2;
 
 public interface AccountsProjection2 {
 
-    int getAccountId();
+    Long getAccountId();
     int getClientId();
     double getBalance();
 }
