+++ src/main/java/com/example/sns/projection/bank2/ExternalTransactionsProjection2.java
@@ -2,7 +2,7 @@ package com.example.sns.projection.bank2;
 
 public interface ExternalTransactionsProjection2 {
 
-    int getexternalTransactionId();
+    Long getexternalTransactionId();
     int getlocalTransactionId();
     String getStatus();
 }
